requirements.txt
appdirs==1.4.3
dj-static==0.0.6
Django==1.8.10
django-braces==1.11.0
django-shortuuidfield==0.1.3
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
shortuuid==0.5.0
six==1.10.0
static3==0.7.0
